refactor(ui): move getSeverityIcon to icon/utils
